Conflicts:
	src/main/webapp/pages/home.html
	src/main/webapp/pages/layout/main.html

+++ src/main/java/com/great/cms/controller/quiz/QuizController.java
@@ -17,12 +17,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.great.cms.controller.utils.IQuizTypeUtil;
 import com.great.cms.controller.utils.QuestionUtil;
 import com.great.cms.controller.utils.QuizTypeUtil;
 import com.great.cms.entity.Question;
 import com.great.cms.entity.Quiz;
 import com.great.cms.entity.Teacher;
 import com.great.cms.entity.Teaches;
+import com.great.cms.enums.QuizStatusType;
 import com.great.cms.security.utils.UserUtil;
 import com.great.cms.service.QuestionService;
 import com.great.cms.service.QuizQuestionService;
@@ -87,6 +89,13 @@ public class QuizController {
 	public String deleteQuiz(@PathVariable Long id,
 			RedirectAttributes redirectAttr) {
 		System.out.println("GET: quiz/delete/{id}" + id);
+		Quiz quiz = quizService.getQuiz(id);
+		QuizStatusType quizStatusType = QuizTypeUtil.getInstance()
+				.getQuizStatusType(quiz, new Date());
+		if (quizStatusType == QuizStatusType.Running) {
+			redirectAttr.addAttribute("id", id);
+			return "redirect:/quiz/view/{id}";
+		}
 		quizService.delete(id);
 		return "redirect:/teacher/quiz/dashboard";
 	}
@@ -99,7 +108,7 @@ public class QuizController {
 		List<Question> assignedQuestions = questionService
 				.findAssignedQuestions(savedQuiz);
 
-		long totalMarks = QuestionUtil.getInstance().getTotalMarks(
+		double totalMarks = QuestionUtil.getInstance().getTotalMarks(
 				assignedQuestions);
 		int totalQuestions = QuestionUtil.getInstance().countTotalQuestions(
 				assignedQuestions);
